Move generators to the top level
