+++ src/generators/batchFileGenerators.ts
@@ -3,9 +3,9 @@ import path from "path";
 import parquet from "@dsnp/parquetjs";
 import { times } from "lodash";
 import { generateBroadcast, generateReply, generateReaction } from "./dsnpGenerators";
-import * as pq from "../../core/batch/parquetSchema";
-import { EthereumAddress } from "../../types/Strings";
-import { BatchBroadcastMessage, BatchReactionMessage, BatchReplyMessage } from "../../core/batch/batchMesssages";
+import * as pq from "../core/batch/parquetSchema";
+import { EthereumAddress } from "../types/Strings";
+import { BatchBroadcastMessage, BatchReactionMessage, BatchReplyMessage } from "../core/batch/batchMesssages";
 import { generateHexString } from "@dsnp/test-generators";
 
 export type DSNPBatchWriteResult = {
