+++ src/generators/index.ts
@@ -1,0 +1,8 @@
+import * as activityPubImport from "./activityPubGenerators";
+export const activityPub = activityPubImport;
+
+import * as batchImport from "./batchFileGenerators";
+export const batch = batchImport;
+
+import * as dsnpImport from "./dsnpGenerators";
+export const dsnp = dsnpImport;
